// client/src/types/ws.ts
export enum WS_MESSAGES {
  CONNECT_ERROR = 'connect_error',
  TOKEN_ERROR = 'token_error',
  LOGIN = 'login',
  LOGIN_ERROR = 'login_error',
  LOGIN_SUCCESS = 'login_success',
  GET_USERS = 'get_users',
  DELETE_USER = 'delete_user',
  DELETE_USER_ERROR = 'delete_user_error',
  DELETE_USER_SUCCESS = 'delete_user_success',
  EDIT_USER = 'edit_user',
  EDIT_USER_ERROR = 'edit_user_error',
  EDIT_USER_SUCCESS = 'edit_user_success',
  CREATE_USER = 'create_user',
  CREATE_USER_ERROR = 'create_user_error',
  CREATE_USER_SUCCESS = 'create_user_success',
  UPLOAD = 'upload',
  UPLOAD_SUCCESS = 'upload_success',
  UPLOAD_ERROR = 'upload_error',
}